Refactor results saving
Improved the environment detection logic to prioritize Kaggle, streamlined the process for saving results in Colab, Kaggle, and local environments, and removed verbose file listing in favor of concise summary output. Added automatic zipping of results and a download link for easier access, and reorganized code for clarity and maintainability.
Add scifact results.